+++ app/com/lynxanalytics/biggraph/controllers/FEAggregations.scala
@@ -24,7 +24,7 @@ object AttributeWithLocalAggregator {
       case "most_common" =>
         AttributeWithLocalAggregator(attr, graph_operations.Aggregator.MostCommon[T]())
       case "count_most_common" =>
-        AttributeWithLocalAggregator(attr, graph_operations.Aggregator.MostCommonCount[T]())
+        AttributeWithLocalAggregator(attr, graph_operations.Aggregator.CountMostCommon[T]())
       case "count_distinct" =>
         AttributeWithLocalAggregator(attr, graph_operations.Aggregator.CountDistinct[T]())
       case "majority_50" =>