+++ test/com/lynxanalytics/biggraph/controllers/GraphDrawingControllerTest.scala
@@ -178,10 +178,27 @@ class GraphDrawingControllerTest extends FunSuite with TestGraphOp {
       mode = "bucketed"
     )
     val res = controller.relativeEdgeDensity(ed, vd, vd)
-    assert(res.edges(0).size == 0.2)
-    assert(res.edges(1).size == 0.1)
-    assert(res.edges(2).size == 0.02)
-    assert(res.edges(3).size == 0.1)
+    val edgeSizes = res.edges.map(_.size)
+    assert(edgeSizes == Seq(0.2, 0.1, 0.02, 0.1))
+  }
+
+  test("relative edge density with zeros") {
+    val ed = EdgeDiagramResponse(
+      srcDiagramId = "vertexDiagramId",
+      dstDiagramId = "vertexDiagramId",
+      srcIdx = 0,
+      dstIdx = 0,
+      edges = Seq(FEEdge(0, 1, 1.0), FEEdge(0, 2, 1.0), FEEdge(1, 2, 1.0), FEEdge(1, 3, 1.0)),
+      layout3D = Map()
+    )
+    val vd = VertexDiagramResponse(
+      diagramId = "vertexDiagramId",
+      vertices = Seq(FEVertex(0.0, 0, 0), FEVertex(5.0, 0, 1), FEVertex(10.0, 1, 0), FEVertex(2.0, 1, 1)),
+      mode = "bucketed"
+    )
+    val res = controller.relativeEdgeDensity(ed, vd, vd)
+    val edgeSizes = res.edges.map(_.size)
+    assert(edgeSizes == Seq(0.0, 0.0, 0.02, 0.1))
   }
 
   test("big bucketed view") {