+++ web/app/scripts/controllers/report-error.js
@@ -5,7 +5,7 @@ angular.module('biggraph').controller('ReportErrorCtrl', function($scope, $modal
   $scope.message = alert.message || '';
   $scope.details = alert.details ? JSON.stringify(alert.details, null, '  ') : undefined;
 
-  var support = 'support@lynxanalytics.freshdesk.com';
+  var support = 'x+170265586669389@mail.asana.com'; // "Support requests" project in Asana.
   var time = alert.time || (new Date().toString());
   var body = $scope.message;
   if ($scope.details) {